requirements.txt
pandas
numpy
scikit-learn
xgboost
tensorflow
Pillow
tqdm
matplotlib
seaborn